Builder properties should use boolean to indicate initialization
Instead of using null to indicate when a builder property is
uninitialized, each property should have two backing fields: a value
holder and a flag to indicate initialization status. The flag can then
be used in all locations to determine if the holder value has been
initialized. This ultimately allows null to be set into nullable
properties.

Fixes #1
Fixes #6